@@ -7,7 +7,7 @@ const client = require("prom-client");
 const register = new client.Registry();
 // Add a default label which is added to all metrics
 register.setDefaultLabels({
-  app: "clocktower-online"
+  app: "clocktower-online",
 });
 
 const PING_INTERVAL = 30000; // 30 seconds
@@ -22,11 +22,9 @@ if (process.env.NODE_ENV !== "development") {
 const server = https.createServer(options);
 const wss = new WebSocket.Server({
   ...(process.env.NODE_ENV === "development" ? { port: 8081 } : { server }),
-  verifyClient: info =>
+  verifyClient: (info) =>
     info.origin &&
-    !!info.origin.match(
-      /^https?:\/\/([^.]+\.github\.io|localhost|clocktower\.online|eddbra1nprivatetownsquare\.xyz)/i
-    )
+    !!info.origin.match(/^https?:\/\/(townsquare.clocktower.guru|localhost)/i),
 });
 
 function noop() {}
@@ -48,14 +46,14 @@ const metrics = {
     help: "Concurrent Players",
     collect() {
       this.set(wss.clients.size);
-    }
+    },
   }),
   channels_concurrent: new client.Gauge({
     name: "channels_concurrent",
     help: "Concurrent Channels",
     collect() {
       this.set(Object.keys(channels).length);
-    }
+    },
   }),
   channels_list: new client.Gauge({
     name: "channel_players",
@@ -66,35 +64,35 @@ const metrics = {
         this.set(
           { name: channel },
           channels[channel].filter(
-            ws =>
+            (ws) =>
               ws &&
               (ws.readyState === WebSocket.OPEN ||
                 ws.readyState === WebSocket.CONNECTING)
           ).length
         );
       }
-    }
+    },
   }),
   messages_incoming: new client.Counter({
     name: "messages_incoming",
-    help: "Incoming messages"
+    help: "Incoming messages",
   }),
   messages_outgoing: new client.Counter({
     name: "messages_outgoing",
-    help: "Outgoing messages"
+    help: "Outgoing messages",
   }),
   connection_terminated_host: new client.Counter({
     name: "connection_terminated_host",
-    help: "Terminated connection due to host already present"
+    help: "Terminated connection due to host already present",
   }),
   connection_terminated_spam: new client.Counter({
     name: "connection_terminated_spam",
-    help: "Terminated connection due to message spam"
+    help: "Terminated connection due to message spam",
   }),
   connection_terminated_timeout: new client.Counter({
     name: "connection_terminated_timeout",
-    help: "Terminated connection due to timeout"
-  })
+    help: "Terminated connection due to timeout",
+  }),
 };
 
 // register metrics
@@ -113,7 +111,7 @@ wss.on("connection", function connection(ws, req) {
     ws.playerId === "host" &&
     channels[ws.channel] &&
     channels[ws.channel].some(
-      client =>
+      (client) =>
         client !== ws &&
         client.readyState === WebSocket.OPEN &&
         client.playerId === "host"
@@ -149,11 +147,7 @@ wss.on("connection", function connection(ws, req) {
       metrics.connection_terminated_spam.inc();
       return;
     }
-    const messageType = data
-      .toLocaleLowerCase()
-      .substr(1)
-      .split(",", 1)
-      .pop();
+    const messageType = data.toLocaleLowerCase().substr(1).split(",", 1).pop();
     switch (messageType) {
       case '"ping"':
         // ping messages will only be sent host -> all or all -> host
@@ -232,7 +226,7 @@ const interval = setInterval(function ping() {
     if (
       !channels[channel].length ||
       !channels[channel].some(
-        ws =>
+        (ws) =>
           ws &&
           (ws.readyState === WebSocket.OPEN ||
             ws.readyState === WebSocket.CONNECTING)
@@ -255,6 +249,6 @@ if (process.env.NODE_ENV !== "development") {
   server.listen(8080);
   server.on("request", (req, res) => {
     res.setHeader("Content-Type", register.contentType);
-    register.metrics().then(out => res.end(out));
+    register.metrics().then((out) => res.end(out));
   });
 }
