@@ -5,11 +5,13 @@
     @close="toggleModal('roles')"
   >
     <h3>Select the characters for {{ nonTravelers }} players:</h3>
+
     <ul class="tokens" v-for="(teamRoles, team) in roleSelection" :key="team">
       <li class="count" :class="[team]">
         {{ teamRoles.reduce((a, { selected }) => a + selected, 0) }} /
         {{ game[nonTravelers - 5][team] }}
       </li>
+
       <li
         v-for="role in teamRoles"
         :class="[role.team, role.selected ? 'selected' : '']"
@@ -17,40 +19,51 @@
         @click="role.selected = role.selected ? 0 : 1"
       >
         <Token :role="role" />
+
         <font-awesome-icon icon="exclamation-triangle" v-if="role.setup" />
+
         <div class="buttons" v-if="allowMultiple">
-          <font-awesome-icon
-            icon="minus-circle"
-            @click.stop="role.selected--"
-          />
+          <div @click.stop="role.selected--">
+            <font-awesome-icon icon="minus-circle" />
+          </div>
+
           <span>{{ role.selected > 1 ? "x" + role.selected : "" }}</span>
-          <font-awesome-icon icon="plus-circle" @click.stop="role.selected++" />
+
+          <div @click.stop="role.selected++">
+            <font-awesome-icon icon="plus-circle" />
+          </div>
         </div>
       </li>
     </ul>
+
     <div class="warning" v-if="hasSelectedSetupRoles">
       <font-awesome-icon icon="exclamation-triangle" />
+
       <span>
         Warning: there are characters selected that modify the game setup! The
         randomizer does not account for these characters.
       </span>
     </div>
+
     <label class="multiple" :class="{ checked: allowMultiple }">
       <font-awesome-icon :icon="allowMultiple ? 'check-square' : 'square'" />
+
       <input type="checkbox" name="allow-multiple" v-model="allowMultiple" />
       Allow duplicate characters
     </label>
+
     <div class="button-group">
       <div
         class="button"
         @click="assignRoles"
         :class="{
-          disabled: selectedRoles > nonTravelers || !selectedRoles
+          disabled: selectedRoles > nonTravelers || !selectedRoles,
         }"
       >
         <font-awesome-icon icon="people-arrows" />
         Assign {{ selectedRoles }} characters randomly
       </div>
+
       <div class="button" @click="selectRandomRoles">
         <font-awesome-icon icon="random" />
         Shuffle characters
@@ -65,53 +78,55 @@ import gameJSON from "./../../game";
 import Token from "./../Token";
 import { mapGetters, mapMutations, mapState } from "vuex";
 
-const randomElement = arr => arr[Math.floor(Math.random() * arr.length)];
+const randomElement = (arr) => arr[Math.floor(Math.random() * arr.length)];
 
 export default {
   components: {
     Token,
-    Modal
+    Modal,
   },
-  data: function() {
+  data: function () {
     return {
       roleSelection: {},
       game: gameJSON,
-      allowMultiple: false
+      allowMultiple: false,
     };
   },
   computed: {
-    selectedRoles: function() {
+    selectedRoles: function () {
       return Object.values(this.roleSelection)
-        .map(roles => roles.reduce((a, { selected }) => a + selected, 0))
+        .map((roles) => roles.reduce((a, { selected }) => a + selected, 0))
         .reduce((a, b) => a + b, 0);
     },
-    hasSelectedSetupRoles: function() {
-      return Object.values(this.roleSelection).some(roles =>
-        roles.some(role => role.selected && role.setup)
+    hasSelectedSetupRoles: function () {
+      return Object.values(this.roleSelection).some((roles) =>
+        roles.some((role) => role.selected && role.setup)
       );
     },
     ...mapState(["roles", "modals"]),
     ...mapState("players", ["players"]),
-    ...mapGetters({ nonTravelers: "players/nonTravelers" })
+    ...mapGetters({ nonTravelers: "players/nonTravelers" }),
   },
   methods: {
     selectRandomRoles() {
       this.roleSelection = {};
-      this.roles.forEach(role => {
+      this.roles.forEach((role) => {
         if (!this.roleSelection[role.team]) {
-          this.$set(this.roleSelection, role.team, []);
+          this.roleSelection[role.team] = [];
+          //this.$set(this.roleSelection, role.team, []);
         }
         this.roleSelection[role.team].push(role);
-        this.$set(role, "selected", 0);
+        role["selected"] = 0;
+        //this.$set(role, "selected", 0);
       });
       delete this.roleSelection["traveler"];
       const playerCount = Math.max(5, this.nonTravelers);
       const composition = this.game[playerCount - 5];
-      Object.keys(composition).forEach(team => {
+      Object.keys(composition).forEach((team) => {
         for (let x = 0; x < composition[team]; x++) {
           if (this.roleSelection[team]) {
             const available = this.roleSelection[team].filter(
-              role => !role.selected
+              (role) => !role.selected
             );
             if (available.length) {
               randomElement(available).selected = 1;
@@ -124,32 +139,32 @@ export default {
       if (this.selectedRoles <= this.nonTravelers && this.selectedRoles) {
         // generate list of selected roles and randomize it
         const roles = Object.values(this.roleSelection)
-          .map(roles =>
+          .map((roles) =>
             roles
               // duplicate roles selected more than once and filter unselected
               .reduce((a, r) => [...a, ...Array(r.selected).fill(r)], [])
           )
           // flatten into a single array
           .reduce((a, b) => [...a, ...b], [])
-          .map(a => [Math.random(), a])
+          .map((a) => [Math.random(), a])
           .sort((a, b) => a[0] - b[0])
-          .map(a => a[1]);
-        this.players.forEach(player => {
+          .map((a) => a[1]);
+        this.players.forEach((player) => {
           if (player.role.team !== "traveler" && roles.length) {
             const value = roles.pop();
             this.$store.commit("players/update", {
               player,
               property: "role",
-              value
+              value,
             });
           }
         });
         this.$store.commit("toggleModal", "roles");
       }
     },
-    ...mapMutations(["toggleModal"])
+    ...mapMutations(["toggleModal"]),
   },
-  mounted: function() {
+  mounted: function () {
     if (!Object.keys(this.roleSelection).length) {
       this.selectRandomRoles();
     }
@@ -157,8 +172,8 @@ export default {
   watch: {
     roles() {
       this.selectRandomRoles();
-    }
-  }
+    },
+  },
 };
 </script>
 
@@ -178,7 +193,7 @@ ul.tokens {
       .buttons {
         display: flex;
       }
-      .fa-exclamation-triangle {
+      .fa-triangle-exclamation {
         display: block;
       }
     }
@@ -201,7 +216,7 @@ ul.tokens {
       transform: scale(1.2);
       z-index: 10;
     }
-    .fa-exclamation-triangle {
+    .fa-triangle-exclamation {
       position: absolute;
       color: red;
       filter: drop-shadow(0 0 3px black) drop-shadow(0 0 3px black);
