@@ -1,5 +1,20 @@
 module.exports = {
-  // if the app is supposed to run on Github Pages in a subfolder, use the following config:
-  // publicPath: process.env.NODE_ENV === "production" ? "/townsquare/" : "/"
-  publicPath: process.env.NODE_ENV === "production" ? "/" : "/"
+  publicPath: process.env.NODE_ENV === "production" ? "/" : "/",
+  chainWebpack: (config) => {
+    config.resolve.alias.set("vue", "@vue/compat");
+
+    config.module
+      .rule("vue")
+      .use("vue-loader")
+      .tap((options) => {
+        return {
+          ...options,
+          compilerOptions: {
+            compatConfig: {
+              MODE: 3,
+            },
+          },
+        };
+      });
+  },
 };
